requirements.txt
beautifulsoup4==4.7.1
pytest==4.1.1
python-decouple==3.1
pytz==2018.9
redis==3.0.1
requests==2.21.0
setuptools==40.6.3
wheel==0.32.3